feat: ajustes de UI, login social, gerenciamento de carrinho e filtro de pesquisa em ativos
- Ajustes na interface para melhor experiência do usuário
- Implementação da funcionalidade de login social
- Adição da funcionalidade de adicionar/remover itens no carrinho
- Integração de filtro de pesquisa para ativos

+++ apps/tech_assets/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from apps.tech_assets.context_processors import get_profile_foto
 from apps.tech_assets.models import Approval, Asset, AssetCart, Cart
-from apps.tech_assets.services import get_user_photo_microsoft, register_logentry
+from apps.tech_assets.services import register_logentry
 from django.contrib.admin.models import CHANGE, DELETION, ADDITION
 from django.shortcuts import get_object_or_404, render, redirect
 from apps.tech_assets.forms import AssetModelForms, LoanForms, AssetForms, MaintenanceForms, \
@@ -11,6 +11,8 @@ from django.urls import resolve
 from django.views.generic import ListView
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
+from django.db.models import Exists, OuterRef, Q
+from django.core.paginator import Paginator
 
 # Create your views here.
 
@@ -18,18 +20,16 @@ from django.contrib.auth.decorators import login_required
 def login(request):
     return render(request, 'shared/login.html')
 
+
 @login_required
 def index(request):
-    print(f'DEBUG :: VIEW :: INDEX :: {request.user}')
     return render(request, 'apps/tech_assets/index.html')
+
 
 @login_required
 def cadastro_fabricante(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = ManufacturerForms
 
@@ -37,9 +37,8 @@ def cadastro_fabricante(request):
         form = ManufacturerForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -48,13 +47,11 @@ def cadastro_fabricante(request):
 
     return render(request, 'apps/tech_assets/cadastro.html', {'form': form, 'url_form': resolve(request.path_info).url_name})
 
+
 @login_required
 def cadastro_modelo(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = AssetModelForms
 
@@ -62,9 +59,8 @@ def cadastro_modelo(request):
         form = AssetModelForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -73,13 +69,11 @@ def cadastro_modelo(request):
 
     return render(request, 'apps/tech_assets/cadastro.html', {'form': form, 'url_form': resolve(request.path_info).url_name})
 
+
 @login_required
 def cadastro_centro_custo(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = CostCenterForms
 
@@ -87,9 +81,8 @@ def cadastro_centro_custo(request):
         form = CostCenterForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -101,11 +94,8 @@ def cadastro_centro_custo(request):
 
 @login_required
 def cadastro_tipo_ativo(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = AssetTypeForms
 
@@ -113,9 +103,8 @@ def cadastro_tipo_ativo(request):
         form = AssetTypeForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -127,11 +116,8 @@ def cadastro_tipo_ativo(request):
 
 @login_required
 def cadastro_local(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = LocationForms
 
@@ -139,9 +125,8 @@ def cadastro_local(request):
         form = LocationForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -153,11 +138,8 @@ def cadastro_local(request):
 
 @login_required
 def cadastro_manutencao(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = MaintenanceForms
 
@@ -165,9 +147,8 @@ def cadastro_manutencao(request):
         form = MaintenanceForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -179,23 +160,17 @@ def cadastro_manutencao(request):
 
 @login_required
 def cadastro_ativo(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = AssetForms
-
-    print(f'DEBUG :: VIEW :: CADASTRO ATIVO :: {form.form_name}')
 
     if request.method == 'POST':
         form = AssetForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -207,11 +182,8 @@ def cadastro_ativo(request):
 
 @login_required
 def novo_emprestimo(request):
-    """
     if not request.user.is_authenticated:
-        print(f'DEBUG :: INDEX :: NÃO TEM USUÁRIO')
-        #messages.error(request, "Usuário não autenticado!")
-        return redirect('login')"""
+        return redirect('login')
 
     form = LoanForms
 
@@ -221,9 +193,8 @@ def novo_emprestimo(request):
         form = LoanForms(request.POST, request.FILES)
 
         if form.is_valid():
-            # register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
+            register_logentry(instance=form.save(usuario=request.user), action=ADDITION, user=request.user)
             # messages.success(request, 'Nova imagem cadastrada na galeria')
-            form.save()
 
             if 'save' in request.POST:
                 return redirect('index')
@@ -235,9 +206,11 @@ def novo_emprestimo(request):
 
 @login_required
 def carrinho(request):
-    user_instance = User.objects.get(username='thiago.leite')
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     # Obtenha o carrinho do usuário atual
-    cart, created = Cart.objects.get_or_create(usuario_sessao=user_instance)
+    cart, created = Cart.objects.get_or_create(usuario_sessao=request.user)
 
     # Recupere os itens do carrinho
     cart_items = AssetCart.objects.filter(carrinho=cart)
@@ -260,9 +233,12 @@ def carrinho(request):
 
 @login_required
 def add_carrinho(request, asset_id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     # Assume que o usuário está autenticado
-    asset = Asset.objects.get(id=asset_id)
-    user_instance = User.objects.get(username='thiago.leite')
+    asset = get_object_or_404(Asset, id=asset_id)
+    user_instance = get_object_or_404(User, username=request.user)
 
     if asset:
         # Verifique se já existe um carrinho para este usuário ou crie um novo
@@ -287,25 +263,49 @@ def add_carrinho(request, asset_id):
 
 @login_required
 def remove_do_carrinho(request, asset_id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     # Assume que o usuário está autenticado
-    asset = Asset.objects.get(id=asset_id)
-    user_instance = User.objects.get(username='thiago.leite')
-
+    asset = get_object_or_404(Asset, id=asset_id)
+    print(f'DEBUG :: ASSET NOME :: ID {asset_id} {asset.nome}')
+    user_instance = get_object_or_404(User, username=request.user)
+    print(f'DEBUG :: USER NOME :: ID {user_instance.id} {user_instance.username}')
+    
     if asset:
-        cart = Cart.objects.get(
-            usuario_sessao=user_instance)
-        print(f'CART ID {cart.id}')
-        if cart:
-            asset_cart = AssetCart.objects.get(ativo=asset, carrinho=cart)
-
-            asset_cart.delete()
-            return redirect('carrinho')
+        try:
+            cart = get_object_or_404(Cart, usuario_sessao=user_instance)
+            if cart:
+                asset_cart = get_object_or_404(AssetCart, ativo=asset, carrinho=cart)
+                if asset_cart:
+                    asset_cart.delete()
+                return redirect('carrinho')
+        except Exception as e:
+            print(f'ERROR :: REMOVE CARRINHO :: {e}')
 
     return redirect('index')
 
+
+
 @login_required
 def ativos(request):
-    print(f'DEBUG :: INDEX :: SESSION USER :: {request.user}')
-    items = Asset.objects.all()
-
-    return render(request, 'apps/tech_assets/ativos.html', {'items': items})
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
+    subquery = AssetCart.objects.filter(ativo_id=OuterRef('pk')).values('pk')
+    assets_in_cart = [asset.id for asset in Asset.objects.filter(Exists(subquery))]
+    
+    query = request.GET.get('q', '')
+    assets = Asset.objects.all()
+    
+    if query:
+        assets = assets.filter(
+            Q(nome__icontains=query) |
+            Q(patrimonio__icontains=query)
+        )
+        
+    paginator = Paginator(assets, 15)
+    page_number = request.GET.get('page')
+    page_obj = paginator.get_page(page_number)
+    
+    return render(request, 'apps/tech_assets/ativos.html', {'assets_in_cart': assets_in_cart, 'page_obj': page_obj, 'query': query})
